Add files via upload
This is a common message sent to everyone. 
As mentioned in the first lab, do not use the same name of the attributes for arguments to the methods. It has an effect of shadowing the “real” attributes. No marks was deducted in case you did so in this particular seminar. If you have any questions about the marking or the seminar in general mail me at ahana.deb@upf.edu. Thank you! :)